Building Damage and Attack
Buildings now take damage and are destroyed once their health is depleted. Their healthbars display correctly. It takes a while for units to attack buildings, especially since units keep spawning from the buildings. This should be fixed once resource management is done. Tried getting the units to run away but it's quite buggy - might remove if it cant be fixed.